refactor(@appcraft/mui): update components folder structure
