+++ libs/mui/src/components/TypeItemDisplay/index.ts
