+++ libs/mui/src/components/common/index.ts
@@ -1,8 +1,9 @@
+export * from './BoolInput';
 export * from './Breadcrumbs';
 export * from './ListPlaceholder';
+export * from './MenuDialog';
+export * from './NumberInput';
 export * from './PlainTextDialog';
-export * from './TypeItemDisplay';
-export * from './TypeItemMixed';
-export * from './TypeItemPure';
 export * from './TypeItemSelection';
+export * from './TypeListSkeleton';
 export * from './WidgetAppBar';
